infeng: suffstat move from __util__.py to infeng.py
    - __util__.py only contains the class suffstat() which is not
      particularly useful outside of the engine() class. To reduce file
      clutter, the __util__.py has been deleted.
